+++ FlateEncoder.cpp
@@ -51,7 +51,7 @@ void FlateEncoder::Begin(){
 
 void FlateEncoder::WriteData(unsigned  char* data, unsigned long length){
 	//write data
-	strm.avail_in = length;
+	strm.avail_in = (uint)length;
 	strm.next_in = data;
 	strm.avail_out = CHUNK;
     strm.next_out = out;
